HTML <input> placeholder Attribute
Definition and Usage
The placeholder attribute specifies a short hint that describes the
expected value of an input field (e.g. a sample value or a short
description of the expected format).

The short hint is displayed in the input field before the user enters a
value.

Note: The placeholder attribute works with the following input types:
text, search, url, tel, email, and password.